layout w sumie gotowy, mozna wiecej linkow dac ew
