+++ client/app/layout.tsx
@@ -4,6 +4,7 @@ import './globals.css';
 import { Toaster } from '@/components/ui/sonner';
 import { ThemeProvider } from '@/components/theme-provider';
 import { SidebarProvider } from '@/components/ui/sidebar';
+import { AppSidebar } from '@/components/app-sidebar';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -35,10 +36,8 @@ export default function RootLayout({
       defaultTheme="system"
       enableSystem
       disableTransitionOnChange>
-      <SidebarProvider>
-        <Toaster/>
-        {children}
-      </SidebarProvider>
+          <Toaster/>
+          {children}
     </ThemeProvider>
     </body>
     </html>
